requirements.txt
python==3.8.15
pytorch==1.12.0
cudatoolkit==10.2.89
torchvision==0.13.0
pandas==1.5.2
numpy==1.23.5
SimpleITK==2.2.1
tqdm==4.64.1
scikit-image==0.19.3
scikit-learn==1.1.1
transformers==4.24.0
matplotlib==3.6.2
statsmodels==0.13.2